+++ src/plugins/autoColumnSize/autoColumnSize.js
@@ -263,13 +263,13 @@ class AutoColumnSize extends BasePlugin {
     });
 
     if (this.ghostTable.columns.length) {
-      this.hot.batch(() => {
+      this.hot.batchExecution(() => {
         this.ghostTable.getWidths((visualColumn, width) => {
           const physicalColumn = this.hot.toPhysicalColumn(visualColumn);
 
           this.columnWidthsMap.setValueAtIndex(physicalColumn, width);
         });
-      });
+      }, true);
 
       this.measuredColumns = columnsRange.to + 1;
 
@@ -314,7 +314,7 @@ class AutoColumnSize extends BasePlugin {
         this.inProgress = false;
 
         // @TODO Should call once per render cycle, currently fired separately in different plugins
-        this.hot.view.wt.wtOverlays.adjustElementsSize();
+        this.hot.view.adjustElementsSize();
       }
     };
 
@@ -507,11 +507,11 @@ class AutoColumnSize extends BasePlugin {
    */
   clearCache(columns = []) {
     if (columns.length) {
-      this.hot.batch(() => {
+      this.hot.batchExecution(() => {
         arrayEach(columns, (physicalIndex) => {
           this.columnWidthsMap.setValueAtIndex(physicalIndex, null);
         });
-      });
+      }, true);
 
     } else {
       this.columnWidthsMap.clear();